refactor(frontend): [#39] reorganize sidebar components and remove unused files
