+++ src/components/layout/collapsible-section.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { Collapsible, CollapsibleContent } from '@/components/ui/collapsible';
-import { SectionHeader } from '../members/newsidebar/section-headers';
+import { SectionHeader } from '../members/sidebar/section-headers';
 
 interface CollapsibleSectionProps {
   title: string;
